test: move functions to utils from db
