+++ tests/integration/polygonParts/helpers/utils.ts
@@ -1,25 +1,161 @@
-import { type PartFeatureProperties } from '@map-colonies/raster-shared';
+/* eslint-disable @typescript-eslint/no-magic-numbers */
+import { faker } from '@faker-js/faker';
+import {
+  CORE_VALIDATIONS,
+  INGESTION_VALIDATIONS,
+  JobTypes,
+  RASTER_PRODUCT_TYPE_LIST,
+  type PartFeatureProperties,
+  type RasterProductTypes,
+} from '@map-colonies/raster-shared';
 import { booleanEqual } from '@turf/boolean-equal';
 import { feature, featureCollection } from '@turf/helpers';
+import { randomPolygon } from '@turf/random';
 import config from 'config';
-import type { Polygon } from 'geojson';
+import type { Feature, Polygon } from 'geojson';
 import { isMatch } from 'lodash';
+import { randexp } from 'randexp';
 import type { ApplicationConfig } from '../../../../src/common/interfaces';
 import { payloadToInsertPartsDataToHistory } from '../../../../src/polygonParts/DAL/utils';
-import type { FindPolygonPartsResponseBody, ValidatePolygonPartsRequestBody } from '../../../../src/polygonParts/controllers/interfaces';
+import type {
+  ExistsRequestBody,
+  FindPolygonPartsResponseBody,
+  ValidatePolygonPartsRequestBody,
+} from '../../../../src/polygonParts/controllers/interfaces';
 import type { EntityIdentifier, PolygonPartsPayload } from '../../../../src/polygonParts/models/interfaces';
 import { INTERNAL_DB_GEOM_PRECISION } from './constants';
-import { type PartialPolygonPartsPayload, generatePolygonPartsPayload as generatePolygonPartsValidationPayload } from './db';
 import type { PolygonPartsRequestSender } from './requestSender';
-import type { ExpectedPostgresResponse, GetEntitiesMetadata } from './types';
+import type { ExpectedPostgresResponse, GetEntitiesMetadata, PartialPolygonPartsPayload, PolygonPartFeature } from './types';
+
+const generateProductId = (): string => randexp(INGESTION_VALIDATIONS.productId.pattern);
+const generateProductType = (): RasterProductTypes => faker.helpers.arrayElement(RASTER_PRODUCT_TYPE_LIST);
 
 const getApplicationConfig = (): ApplicationConfig => config.get<ApplicationConfig>('application');
 
-export type GeneratePolygonPartsPayloadOverrides = Partial<Omit<ValidatePolygonPartsRequestBody, 'partsData'>> & {
-  partsData?: ValidatePolygonPartsRequestBody['partsData'];
-  geometry?: ValidatePolygonPartsRequestBody['partsData']['features'][number]['geometry'];
-  features?: ValidatePolygonPartsRequestBody['partsData']['features'];
+export const generateFeatureId = (): NonNullable<Feature['id']> => {
+  return faker.helpers.arrayElement([faker.number.float({ max: Number.MAX_VALUE }), faker.string.uuid(), faker.string.alphanumeric({ length: 20 })]);
 };
+
+export const generateResolutionDegree = (): PolygonPartFeature['properties']['resolutionDegree'] =>
+  faker.number.float(CORE_VALIDATIONS.resolutionDeg);
+
+export const generatePolygon = (
+  options: Parameters<typeof randomPolygon>[1] = {
+    bbox: [-170, -80, 170, 80],
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    max_radial_length: faker.number.float({ min: Number.EPSILON, max: 10 }),
+  } // polygon maximum extent cannot exceed [-180,-90,180,90]
+): Polygon => {
+  return randomPolygon(1, options).features[0].geometry;
+};
+
+export const generatePolygonPart = (): PolygonPartFeature => {
+  const date1 = faker.date.past();
+  const date2 = faker.date.past();
+  const [dateOlder, dateRecent] = date1 < date2 ? [date1, date2] : [date2, date1];
+  return {
+    type: 'Feature',
+    id: generateFeatureId(),
+    geometry: generatePolygon(),
+    properties: {
+      id: faker.string.uuid(),
+      horizontalAccuracyCE90: faker.number.float(INGESTION_VALIDATIONS.horizontalAccuracyCE90),
+      imagingTimeBeginUTC: dateOlder,
+      imagingTimeEndUTC: dateRecent,
+      resolutionDegree: generateResolutionDegree(),
+      resolutionMeter: faker.number.float(INGESTION_VALIDATIONS.resolutionMeter),
+      sensors: faker.helpers.multiple(
+        () => {
+          return faker.word.words();
+        },
+        { count: { min: 1, max: 3 } }
+      ),
+      sourceName: faker.word.words().replace(' ', '_'),
+      sourceResolutionMeter: faker.number.float(INGESTION_VALIDATIONS.resolutionMeter),
+      cities: faker.helpers.maybe(() => {
+        return faker.helpers.multiple(
+          () => {
+            return faker.word.words();
+          },
+          { count: { min: 1, max: 3 } }
+        );
+      }),
+      countries: faker.helpers.maybe(() => {
+        return faker.helpers.multiple(
+          () => {
+            return faker.word.words();
+          },
+          { count: { min: 1, max: 3 } }
+        );
+      }),
+      description: faker.helpers.maybe(() => faker.word.words({ count: { min: 0, max: 10 } })),
+      sourceId: faker.helpers.maybe(() => faker.word.words()),
+    },
+  };
+};
+
+export const generateExistsPayload = (): ExistsRequestBody => {
+  return {
+    productId: generateProductId(),
+    productType: generateProductType(),
+  };
+};
+
+export function generatePolygonPartsPayload(partsCount?: number): PolygonPartsPayload;
+export function generatePolygonPartsPayload(template?: PartialPolygonPartsPayload): PolygonPartsPayload;
+export function generatePolygonPartsPayload(input?: number | PartialPolygonPartsPayload): PolygonPartsPayload {
+  const layerMetadata = {
+    catalogId: faker.string.uuid(),
+    productId: generateProductId(),
+    productType: generateProductType(),
+    productVersion: randexp(INGESTION_VALIDATIONS.productVersion.pattern),
+    jobType: faker.helpers.arrayElement([JobTypes.Ingestion_New, JobTypes.Ingestion_Update, JobTypes.Ingestion_Swap_Update]),
+  } satisfies Omit<PolygonPartsPayload, 'partsData'>;
+
+  if (typeof input === 'number' || input === undefined) {
+    const partsCount = input ?? 1;
+    return {
+      ...layerMetadata,
+      partsData: {
+        type: 'FeatureCollection',
+        features: Array.from({ length: partsCount }, generatePolygonPart),
+      },
+    };
+  }
+
+  const { partsData: templatePartsData, ...templateLayerMetadata } = structuredClone(input);
+  const features = templatePartsData?.features;
+  const featureCount = features?.length ?? 1;
+
+  return {
+    ...layerMetadata,
+    ...templateLayerMetadata,
+    partsData: {
+      type: 'FeatureCollection',
+      features: Array.from({ length: featureCount }, generatePolygonPart).map((partData, index) => {
+        const templateFeature = features?.[index];
+
+        if (!templateFeature) {
+          return partData;
+        }
+
+        const feature = {
+          ...partData,
+          ...(templateFeature.id !== undefined && { id: templateFeature.id }),
+          ...(templateFeature.geometry !== undefined && { geometry: templateFeature.geometry }),
+          ...(templateFeature.properties !== undefined && {
+            properties: {
+              ...partData.properties,
+              ...templateFeature.properties,
+            },
+          }),
+        };
+
+        return feature;
+      }),
+    },
+  };
+}
 
 export const allFindFeaturesEqual = <T extends FindPolygonPartsResponseBody<ShouldClip>['features'][number], ShouldClip extends boolean = boolean>(
   expectedGeometries: Polygon[],
@@ -49,7 +185,7 @@ export const insertInitialPolygonParts = async ({
   requestSender: PolygonPartsRequestSender;
   getEntitiesMetadata: GetEntitiesMetadata;
 }): Promise<{ entityIdentifier: EntityIdentifier; maxResolutionDegree: number; minResolutionDegree: number }> => {
-  const { partsData, ...layerMetadata } = generatePolygonPartsValidationPayload(input);
+  const { partsData, ...layerMetadata } = generatePolygonPartsPayload(input);
   const validatePartsData = structuredClone(partsData);
   const validateRequest: ValidatePolygonPartsRequestBody = { ...layerMetadata, jobType: 'Ingestion_New', partsData: validatePartsData };
   await requestSender.validatePolygonParts(validateRequest);
